+++ libgen_to_txt/naive/other.py
